convert LI.status to long

+++ app/src/main/java/com/trulden/friends/database/FriendsDao.java
@@ -165,7 +165,7 @@ public interface FriendsDao {
         "  t2\n" +
         "  ON (t1.interId = t2.interactionId); "
     )
-    void recalcLastInteraction(long typeId, long friendId, int status);
+    void recalcLastInteraction(long typeId, long friendId, long status);
 
     @Query("SELECT status FROM last_interaction_table " +
             "WHERE typeId = :typeId " +
